Finish sample create lottery in admin controller

+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/Request/CreateLotteryRequest.java
@@ -7,6 +7,6 @@ import lombok.Data;
 public class CreateLotteryRequest {
 
     public String ticket;
-    public int price;
+    public Long price;
     public int amount;
 }
